+++ frontend/src/_hooks/use-mount.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const useMounted = () => {
+  const [mounted, setMounted] = React.useState(false);
+  React.useEffect(() => {
+    setMounted(true);
+  }, []);
+  return mounted;
+};
